// category/src/modules/grpc/proto/category.pb.ts
export interface GetCatConPriceRangeRequest {
  variantId: string;
  conditionId: string;
  catConditionQuality: string;
}

export interface GetCatConPriceRangeResponse {
  priceRange: PriceNudge | undefined;
}

export interface PriceNudge {
  min: number;
  max: number;
}

export interface GetProductCatConRequest {
  id: string;
  variantId: string;
  sellPrice: number;
}

export interface GetProductCatConResponse {
  condition: Condition | undefined;
  priceQuality: PriceQuality | undefined;
}

export interface Banner {
  lang: string;
  url: string;
  source: string;
}

export interface PriceQuality {
  name: string;
}

export interface Condition {
  id: string;
  name: string;
  nameAr: string;
  labelColor: string;
  textColor: string;
  banners: Banner[];
  positionEn: number;
  positionAr: number;
}

export interface CategoryIds {
  ids: string[];
}

export interface GetCategoriesRequest {
  limit: number;
  offset: number;
  type: string;
  categories?: CategoryIds | undefined;
}

export interface GetCategoryByNameRequest {
  name: string;
}

export interface Category {
  id: string;
  name: string;
  nameAr: string;
  status: string;
  type: string;
  parentId: string;
  icon: Icon | undefined;
  position: number;
}

export interface GetCategoriesResponse {
  categories: Category[];
}

export interface CreateCategoryRequest {
  name: string;
  nameAr: string;
  position: number;
  status: string;
  type: string;
  parentId: string;
  icons: Icon[];
  id: string;
  maxPercentage?: number | undefined;
  currentPrice?: number | undefined;
}

export interface Icon {
  type: string;
  url: string;
  source: string;
}

export interface CreateCategoryResponse {
}

export interface GetAttributesRequest {
  size: number;
  page: number;
  search: string;
  optionsIncluded?: boolean | undefined;
}

export interface Option {
  id: string;
  nameEn: string;
  nameAr: string;
  status: string;
  positionAr: number;
  positionEn: number;
}

export interface Attribute {
  id: string;
  nameEn: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface UpdateAttributeRequest {
  id: string;
  nameEn: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface UpdateAttributeResponse {
  id: string;
  name: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface GetAttributesResponse {
  attributes: Attribute[];
  total: number;
}

export interface DeleteAttributeRequest {
  id: string;
}

export interface DeleteAttributeResponse {
}

export interface GetAttributeRequest {
  id: string;
}

export interface GetAttributeResponse {
  attribute: Attribute | undefined;
}

export interface CreateAttributeRequest {
  name: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface CreateAttributeResponse {
  id: string;
  name: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface GetConditionsRequest {
  ids: string[];
}

export interface GetConditionsResponse {
  conditions: Condition[];
}

export interface GetMultipleAttributeRequest {
  ids: string[];
}

export interface GetMultipleAttributeResponse {
  attributes: Attribute[];
}
// ...
export const CATEGORY_SERVICE_NAME = "CategoryService";